fix(rpc): degrade to unproven when there is nothing to anchor to
`latest_commit_proof` returned `Ok(None)` only for a `LastExecuted` row stored at
the queried epoch with height zero. Both real "nothing committed here" cases -
no row at all, and a row belonging to a different epoch - come back from
`last_executed_get` as `NotFound` and were propagated as errors. So the window
right after an epoch change, and a node that has not started consensus since
restarting, failed the request instead of answering unproven. `.optional()?`
gives both callers the `None` they were already written against.

A substate the anchor's shard group does not cover is now `Ok(None)` from
`SubstateProofGenerator::generate` rather than an error, alongside the existing
"shard has no committed state" case: both mean this state cannot speak for that
substate, neither means the read failed. The batch reports it as missing instead
of failing the other forty-nine, and `get_substate` answers unproven instead of
erroring. The up-front membership check still rejects ids this node does not
store, which is the case that is actually a client error.

`generate_substate_proof` goes with it - `attach_substate_proof` was its last
caller and now needs the `Option` the generator returns.
