+++ crates/state_store_rocksdb/tests/helpers.rs
@@ -1,7 +1,7 @@
 //   Copyright 2026 The Tari Project
 //   SPDX-License-Identifier: BSD-3-Clause
 
-use std::{io::Write, ops::Deref};
+use std::{collections::BTreeMap, io::Write, ops::Deref};
 
 use rand::{Rng, RngExt};
 use tari_bor::cbor;
@@ -23,6 +23,8 @@ use tari_ootle_common_types::{
     shard::Shard,
 };
 use tari_ootle_storage::{
+    ShardScopedTreeStoreWriter,
+    StateStore,
     StateStoreReadTransaction,
     StateStoreWriteTransaction,
     consensus_models::{
@@ -43,7 +45,7 @@ use tari_ootle_storage::{
 use tari_ootle_transaction::{Network, TransactionId};
 use tari_sidechain::{CommitProofElement, QuorumDecision, SidechainBlockCommitProof, SidechainBlockHeader};
 use tari_state_store_rocksdb::{DatabaseOptions, RocksDbStateStore};
-use tari_state_tree::Version;
+use tari_state_tree::{SpreadPrefixStateTree, SubstateTreeChange, Version};
 use tari_template_lib::types::{
     ComponentAddress,
     ComponentKey,
@@ -469,4 +471,38 @@ pub fn create_foreign_proposal(parent_id: BlockId, epoch: Epoch) -> ForeignPropo
     });
 
     ForeignProposalRecord::new(ForeignProposal::new(commit_proof, BlockPledge::default()))
+}
+
+/// The state-tree version the substates committed by [`commit_substates`] land at.
+pub const PROOF_TEST_TREE_VERSION: Version = 1;
+
+/// Commits `substates` to the store and to their shards' state trees, as a validator does when a
+/// block commits, so that proofs can be generated against the resulting shard-group root.
+pub fn commit_substates(db: &impl StateStore, substates: &[SubstateRecord]) {
+    let mut by_shard: BTreeMap<Shard, Vec<&SubstateRecord>> = BTreeMap::new();
+    for substate in substates {
+        by_shard.entry(substate.created().in_shard).or_default().push(substate);
+    }
+
+    let mut tx = db.create_write_tx().unwrap();
+    Block::zero_block(NETWORK, num_preshards()).insert(&mut tx).unwrap();
+
+    for (shard, substates) in &by_shard {
+        let changes = substates.iter().map(|s| SubstateTreeChange::Up {
+            id: s.to_versioned_substate_id(),
+            value_hash: *s.state_hash(),
+        });
+        {
+            let mut store = ShardScopedTreeStoreWriter::new(&mut tx, *shard);
+            SpreadPrefixStateTree::new(&mut store)
+                .batch_put_substate_changes(None, PROOF_TEST_TREE_VERSION, changes)
+                .unwrap();
+        }
+        tx.state_tree_shard_versions_set(*shard, PROOF_TEST_TREE_VERSION)
+            .unwrap();
+    }
+
+    tx.substates_commit_batch(create_substate_update_batch(Epoch::zero(), substates))
+        .unwrap();
+    tx.commit().unwrap();
 }
